Add recoil for state management
Also - begin setting up a basic structure for the current project and centralized way of knowing where to navigate and what is the current global state. I needed this to make the routing work. I am not sure if there is a smarter way thoug... must be as I feel like this is a bit wrong.